requirements.txt
line-bot-sdk
flask
requests
BeautifulSoup4